re-organizing code and revising impl notes
