+++ Sources/Formic/ResourceTypes/OperatingSystem.swift
